Refactor analytics utilities into modular files
